refactor(server): deprecate card creation API
