+++ functions/src/handler/add_credit_card.ts
@@ -1,7 +1,13 @@
-import { noAuthMsg } from "../config/consts";
+import { noAuthMsg, requestDataIncomplete } from "../config/consts";
 import { CreditCardIdConflictError } from "../config/errors";
 import { FunctionContext } from "../config/typedefs";
-import { CreditCardCreationRequest } from "../interfaces/interfaces";
+import {
+  CreditCardCreationRequest,
+  Promotion,
+  CreditCard,
+} from "../interfaces/interfaces";
+import { setPromotion } from "./setters/set_promotion";
+import { setCreditCard } from "./setters/set_credit_card";
 
 async function addCreditCardHandler(
   data: CreditCardCreationRequest,
@@ -9,25 +15,35 @@ async function addCreditCardHandler(
   provider
 ) {
   if (context.authenticated) {
-    const userUid: string = context.uid;
-    const userRef = provider.getUserRef(userUid);
-    const cardRef = userRef.collection("cards").doc(data.cardData?.id);
-    const cardSnap = await cardRef.get();
-    if (cardSnap.exists) {
-      throw CreditCardIdConflictError;
+    if (data.cardData) {
+      const userId: string = context.uid;
+      const userRef = provider.getUserRef(userId);
+      const cardRef = userRef.collection("cards").doc(data.cardData.id);
+      const cardSnap = await cardRef.get();
+      if (cardSnap.exists) {
+        throw CreditCardIdConflictError;
+      } else {
+        await setCreditCard(userId, CreditCard.create(data.cardData), provider);
+        if (data.cardData.promotions) {
+          for (const promo of data.cardData.promotions) {
+            // The loggings are here to check if there is a
+            // race condition. It's not likely, but good to
+            // know.
+            console.log("start adding " + promo.id);
+            if (data.cardData.id) {
+              await setPromotion(
+                userId,
+                data.cardData.id,
+                Promotion.create(promo),
+                provider
+              );
+            }
+            console.log("finished adding " + promo.id);
+          }
+        }
+      }
     } else {
-      const cardName: string =
-        data.cardData && data.cardData.displayName
-          ? data.cardData.displayName
-          : "na";
-      const officialUrl: string = data.cardData?.officialUrl
-        ? data.cardData?.officialUrl
-        : "na";
-      const status = await cardRef.set({
-        card_name: cardName,
-        official_url: officialUrl,
-      });
-      return status;
+      throw requestDataIncomplete;
     }
   } else {
     throw noAuthMsg;
